+++ ethereum-types/src/uint.rs
@@ -8,17 +8,19 @@
 
 #[cfg(feature = "codec")]
 use impl_codec::impl_uint_codec;
+#[cfg(feature = "rlp")]
 use impl_rlp::impl_uint_rlp;
 #[cfg(feature = "impl-serde")]
 use impl_serde::impl_uint_serde;
 use uint_crate::*;
 
-pub use uint_crate::FromDecStrErr;
+pub use uint_crate::{FromDecStrErr, FromStrRadixErr, FromStrRadixErrKind};
 
 construct_uint! {
 	/// Unsigned 64-bit integer.
 	pub struct U64(1);
 }
+#[cfg(feature = "rlp")]
 impl_uint_rlp!(U64, 1);
 #[cfg(feature = "impl-serde")]
 impl_uint_serde!(U64, 1);