+++ ethbloom/CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog].
 
 ## [Unreleased]
 
+## [0.11.0] - 2021-01-27
+### Breaking
+- Updated `impl-codec` to 0.5. [#510](https://github.com/paritytech/parity-common/pull/510)
+
+### Potentially-breaking
+- `serialize` feature no longer pulls `std`. [#503](https://github.com/paritytech/parity-common/pull/503)
+
+## [0.10.0] - 2021-01-05
+### Breaking
+- Updated `rlp` to 0.5. [#463](https://github.com/paritytech/parity-common/pull/463)
+
 ## [0.9.2] - 2020-05-18
 - Added `codec` feature. [#393](https://github.com/paritytech/parity-common/pull/393)