+++ keccak-hash/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog].
 
 ## [Unreleased]
 
+## [0.7.0] - 2021-01-27
+### Breaking
+- Updated `primitive-types` to 0.9. [#510](https://github.com/paritytech/parity-common/pull/510)
+
+## [0.6.0] - 2021-01-05
+### Breaking
+- Updated `primitive-types` to 0.8. [#463](https://github.com/paritytech/parity-common/pull/463)
+
 ## [0.5.1] - 2020-04-10
 - Added `keccak256_range` and `keccak512_range` functions. [#370](https://github.com/paritytech/parity-common/pull/370)